+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # 2.x branch
 ## 2.26 branch
 ### 2.26.1
+* the length of the tags has been increased to 40 characters;
+* added `VersionUpdatesCommand`, that performs some updates to the database or
+    files needed for versioning;
 * fixed a bug when loading plugins, when the `Asset` plugin does not exist.
 
 ### 2.26.0